// abrpoint.client/src/models/PosteHoraire.ts
export interface PosteHoraire {
  codposte: string;
  soccod: string;
  ordre?: number;
  catcod?: string;
  catdu?: Date | null;
  catau?: Date | null;
  catfixe?: string;
  catlib?: string;
  cathsup?: string;
  catperiode?: string;
  libposte?: string;

  avantent?: number;
  apresent?: number;
  avantsort?: number;
  apressort?: number;
  retsanc?: number;
  retmin?: number;
  retsancam?: number;
  retminam?: number;
  avabon?: number;
  avamn?: number;
  avabonam?: number;
  avamnam?: number;

  lunhdmat?: string;
  lunhfmat?: string;
  lunhdam?: string;
  lunhfam?: string;
  lunrepos?: string;
  lunrepas?: number;

  marhdmat?: string;
  marhfmat?: string;
  marhdam?: string;
  marhfam?: string;
  marrepos?: string;
  marrepas?: number;

  merhdmat?: string;
  merhfmat?: string;
  merhdam?: string;
  merhfam?: string;
  merrepos?: string;
  merrepas?: number;

  jeuhdmat?: string;
  jeuhfmat?: string;
  jeuhdam?: string;
  jeuhfam?: string;
  jeurepos?: string;
  jeurepas?: number;

  venhdmat?: string;
  venhfmat?: string;
  venhdam?: string;
  venhfam?: string;
  venrepos?: string;
  venrepas?: number;

  samhdmat?: string;
  samhfmat?: string;
  samhdam?: string;
  samhfam?: string;
  samrepos?: string;
  samrepas?: number;

  dimhdmat?: string;
  dimhfmat?: string;
  dimhdam?: string;
  dimhfam?: string;
  dimrepos?: string;
  dimrepas?: number;

  lunhdrep?: string;
  lunhfrep?: string;
  marhdrep?: string;
  marhfrep?: string;
  merhdrep?: string;
  merhfrep?: string;
  jeuhdrep?: string;
  jeuhfrep?: string;
  venhdrep?: string;
  venhfrep?: string;
  samhdrep?: string;
  samhfrep?: string;
  dimhdrep?: string;
  dimhfrep?: string;

  lunhdematin?: string;
  lunhfematin?: string;
  marhdematin?: string;
  marhfematin?: string;
  merhdematin?: string;
  merhfematin?: string;
  jeuhdematin?: string;
  jeuhfematin?: string;
  venhdematin?: string;
  venhfematin?: string;
  samhdematin?: string;
  samhfematin?: string;
  dimhdematin?: string;
  dimhfematin?: string;

  lunhdeamidi?: string;
  lunhfeamidi?: string;
  marhdeamidi?: string;
  marhfeamidi?: string;
  merhdeamidi?: string;
  merhfeamidi?: string;
  jeuhdeamidi?: string;
  jeuhfeamidi?: string;
  venhdeamidi?: string;
  venhfeamidi?: string;
  samhdeamidi?: string;
  samhfeamidi?: string;
  dimhdeamidi?: string;
  dimhfeamidi?: string;

  arrondi?: number;
  arrhsup?: number;
  arrhsortie?: number;
  arrhsmajore?: number;
  arrhentree?: number;
  arrhemajore?: number;

  maxhrelun?: string;
  maxhremar?: string;
  maxhremer?: string;
  maxhreven?: string;
  maxhresam?: string;
  maxhredim?: string;
  lundouche?: number;
  mardouche?: number;
  merdouche?: number;
  jeudouche?: number;
  vendouche?: number;
  samdouche?: number;
  dimdouche?: number;
  minhjourlun?: number;
  minhjourmar?: number;
  minhjourmer?: number;
  minhjourjeu?: number;
  minhjourven?: number;
  minhjoursam?: number;
  minhjourdim?: number;
  maxhrejeu?: number;
  Dimhdeamidi?: number;
  Dimhfeamidi?: number;
  minhdemijourlun?: number;
  minhdemijourmar?: number;
  minhdemijourmer?: number;
  minhdemijourjeu?: number;
  minhdemijourven?: number;
  minhdemijoursam?: number;
  minhdemijourdim?: number;
  
  catsem2?: string;
  catsem3?: string;
  catsem4?: string;
  catsem5?: string;
  catsem6?: string;
}